Refonte catégories - sous catégories (refs #146) @80% : Fixing modification annonce

+++ batimen-dto/src/main/java/fr/batimen/dto/constant/Categorie.java
@@ -5,11 +5,11 @@ package fr.batimen.dto.constant;
  */
 public class Categorie {
 
-    public static final Integer ELECTRICITE_CODE = 0;
-    public static final Integer PLOMBERIE_CODE = 1;
-    public static final Integer ESPACE_VERT_CODE = 2;
-    public static final Integer DECORATION_MACONNERIE_CODE = 3;
-    public static final Integer MENUISERIE_CODE = 4;
+    public static final Short ELECTRICITE_CODE = 0;
+    public static final Short PLOMBERIE_CODE = 1;
+    public static final Short ESPACE_VERT_CODE = 2;
+    public static final Short DECORATION_MACONNERIE_CODE = 3;
+    public static final Short MENUISERIE_CODE = 4;
 
     public static final String ELECTRICITE_NAME = "Electricité";
     public static final String PLOMBERIE_NAME = "Plomberie";
@@ -36,13 +36,13 @@ public class Categorie {
     }
 
     public static String getIcon(Short codeCategorieMetier) {
-        if (codeCategorieMetier.equals(ELECTRICITE_CODE.shortValue())) {
+        if (codeCategorieMetier.equals(ELECTRICITE_CODE)) {
             return "icon-Lightning";
-        } else if (codeCategorieMetier.equals(PLOMBERIE_CODE.shortValue())) {
+        } else if (codeCategorieMetier.equals(PLOMBERIE_CODE)) {
             return "icons8-plumbing";
-        } else if (codeCategorieMetier.equals(ESPACE_VERT_CODE.shortValue())) {
+        } else if (codeCategorieMetier.equals(ESPACE_VERT_CODE)) {
             return "icons8-garden-shears";
-        } else if (codeCategorieMetier.equals(MENUISERIE_CODE.shortValue())) {
+        } else if (codeCategorieMetier.equals(MENUISERIE_CODE)) {
             return "icon-Forrst";
         } else {
             return "icon-Tool";
